apps/api/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.40
alembic==1.15.2
psycopg2-binary==2.9.10
pydantic==2.11.1
pydantic-settings==2.8.1
PyJWT[crypto]==2.9.0
httpx==0.28.1
redis==5.2.1
python-multipart==0.0.20
